+++ packages/typespec-ts/src/modular/helpers/clientHelpers.ts
@@ -20,6 +20,8 @@ import { SdkContext } from "../../utils/interfaces.js";
 import { getClassicalClientName } from "./namingHelpers.js";
 import { getTypeExpression } from "../type-expressions/get-type-expression.js";
 import { isCredentialType } from "./typeHelpers.js";
+import { CloudSettingHelpers } from "../static-helpers-metadata.js";
+import { resolveReference } from "../../framework/reference.js";
 
 interface ClientParameterOptions {
   onClientOnly?: boolean;
@@ -172,7 +174,10 @@ export function buildGetClientEndpointParam(
 ): string {
   let coreEndpointParam = "";
   if (dpgContext.rlcOptions?.flavor === "azure") {
-    coreEndpointParam = `options.endpoint ?? options.baseUrl`;
+    const cloudSettingSuffix = dpgContext.arm
+      ? ` ?? ${resolveReference(CloudSettingHelpers.getArmEndpoint)}(options.cloudSetting)`
+      : "";
+    coreEndpointParam = `options.endpoint${cloudSettingSuffix}`;
   } else {
     // unbranded does not have the deprecated baseUrl parameter
     coreEndpointParam = `options.endpoint`;