+++ packages/typespec-ts/src/utils/operationUtil.ts
@@ -22,8 +22,6 @@ import {
   getHttpOperationWithCache,
   getWireName,
   InitializedByFlags,
-  listOperationGroups,
-  listOperationsInOperationGroup,
   SdkBodyParameter,
   SdkClient,
   SdkClientType,
@@ -34,7 +32,14 @@ import {
   SdkServiceMethod,
   SdkServiceOperation
 } from "@azure-tools/typespec-client-generator-core";
-import { Model, Operation, Program, Type } from "@typespec/compiler";
+import {
+  isList,
+  Model,
+  NoTarget,
+  Operation,
+  Program,
+  Type
+} from "@typespec/compiler";
 import {
   HttpOperation,
   HttpOperationParameter,
@@ -47,6 +52,9 @@ import { isByteOrByteUnion } from "./modelUtils.js";
 import { getOperationNamespaceInterfaceName } from "./namespaceUtils.js";
 import { resolveReference } from "../framework/reference.js";
 import { SerializationHelpers } from "../modular/static-helpers-metadata.js";
+import { listOperationsUnderRLCClient } from "./clientUtils.js";
+import { $ } from "@typespec/compiler/typekit";
+import { reportDiagnostic } from "../lib.js";
 
 // Sorts the responses by status code
 export function sortedOperationResponses(responses: HttpOperationResponse[]) {
@@ -313,9 +321,8 @@ export function hasPollingOperations(
   dpgContext: SdkContext
 ) {
   const program = dpgContext.program;
-  const clientOperations = listOperationsInOperationGroup(dpgContext, client);
-  for (const clientOp of clientOperations) {
-    const route = getHttpOperationWithCache(dpgContext, clientOp);
+  for (const op of listOperationsUnderRLCClient(client)) {
+    const route = getHttpOperationWithCache(dpgContext, op);
     // ignore overload base operation
     if (route.overloads && route.overloads?.length > 0) {
       continue;
@@ -324,64 +331,90 @@ export function hasPollingOperations(
       return true;
     }
   }
-  const operationGroups = listOperationGroups(dpgContext, client, true);
-  for (const operationGroup of operationGroups) {
-    const operations = listOperationsInOperationGroup(
-      dpgContext,
-      operationGroup
-    );
-    for (const op of operations) {
-      const route = getHttpOperationWithCache(dpgContext, op);
-      // ignore overload base operation
-      if (route.overloads && route.overloads?.length > 0) {
-        continue;
-      }
-      if (isLongRunningOperation(program, route)) {
-        return true;
-      }
-    }
-  }
   return false;
 }
 
-export function isPagingOperation(program: Program, operation: HttpOperation) {
-  for (const response of operation.responses) {
-    const paged = extractPagedMetadataNested(program, response.type as Model);
-    if (paged) {
-      return true;
+export interface PageDetails {
+  nextLinkNames: string[];
+  itemNames: string[];
+}
+
+export function extractPageDetails(
+  program: Program,
+  operation: HttpOperation
+): PageDetails | undefined {
+  if (isList(program, operation.operation)) {
+    // If the operation is a list, we don't need to extract paging details.
+    const metadata = $(program).operation.getPagingMetadata(
+      operation.operation
+    );
+    if (metadata === undefined) {
+      // would fallback to default paging metadata
+      reportDiagnostic(program, {
+        code: "no-paging-items-defined",
+        format: {
+          operationName: operation.operation.name
+        },
+        target: NoTarget
+      });
+    }
+    const nextLinkPath = metadata?.output.nextLink?.path;
+    const itemNamePath = metadata?.output.pageItems?.path;
+    if (
+      (nextLinkPath && nextLinkPath?.length > 1) ||
+      (itemNamePath && itemNamePath?.length > 1)
+    ) {
+      // Any cases with nested nextLink or itemName are not supported
+      reportDiagnostic(program, {
+        code: "un-supported-paging-cases",
+        format: {
+          operationName: operation.operation.name
+        },
+        target: NoTarget
+      });
+      // these paging information will be ignored
+      // and the operation will be treated as a non-paging operation.
+      return undefined;
+    }
+    const nextLinkNames =
+      nextLinkPath?.map((prop) => prop.name).join(".") ?? "nextLink";
+    const itemNames =
+      itemNamePath?.map((prop) => prop.name).join(".") ?? "value";
+    return {
+      nextLinkNames: [nextLinkNames],
+      itemNames: [itemNames]
+    };
+  } else {
+    // TODO: remember to remove this once Azure Paging is removed.
+    for (const response of operation.responses) {
+      const paged = extractPagedMetadataNested(program, response.type as Model);
+      if (paged) {
+        const nextLinkName = parseNextLinkName(paged) ?? "nextLink";
+        const itemName = parseItemName(paged) ?? "value";
+        return {
+          nextLinkNames: [nextLinkName],
+          itemNames: [itemName]
+        };
+      }
     }
   }
-  return false;
+  return undefined;
+}
+
+export function isPagingOperation(program: Program, operation: HttpOperation) {
+  return extractPageDetails(program, operation) !== undefined;
 }
 
 export function hasPagingOperations(client: SdkClient, dpgContext: SdkContext) {
   const program = dpgContext.program;
-  const clientOperations = listOperationsInOperationGroup(dpgContext, client);
-  for (const clientOp of clientOperations) {
-    const route = getHttpOperationWithCache(dpgContext, clientOp);
+  for (const op of listOperationsUnderRLCClient(client)) {
+    const route = getHttpOperationWithCache(dpgContext, op);
     // ignore overload base operation
     if (route.overloads && route.overloads?.length > 0) {
       continue;
     }
     if (isPagingOperation(program, route)) {
       return true;
-    }
-  }
-  const operationGroups = listOperationGroups(dpgContext, client, true);
-  for (const operationGroup of operationGroups) {
-    const operations = listOperationsInOperationGroup(
-      dpgContext,
-      operationGroup
-    );
-    for (const op of operations) {
-      const route = getHttpOperationWithCache(dpgContext, op);
-      // ignore overload base operation
-      if (route.overloads && route.overloads?.length > 0) {
-        continue;
-      }
-      if (isPagingOperation(program, route)) {
-        return true;
-      }
     }
   }
   return false;